[vpr][route][crr] set NodeId to size_T

+++ vpr/src/route/rr_graph_generation/tileable_rr_graph/crr_generator/crr_common.h
@@ -46,7 +46,7 @@ class ThreadPool;
 class RRGraphGenerator;
 
 // Type aliases for convenience
-using NodeId = int32_t;
+using NodeId = size_t;
 using SwitchId = int32_t;
 using SegmentId = int32_t;
 using Coordinate = int32_t;
